move shared functions to lapis
